:truck: Rename and order Auth module files and folders for more clarity
