+++ src/modules/Auth/components/AccountsManagerScreen/styles.js
