+++ src/modules/Auth/components/AuthMethod/__fixtures__/account.js
