+++ src/modules/Auth/components/AuthMethod/__fixtures__/mockEncryptedAccount.js
