+++ src/modules/Auth/components/AuthMethod/styles.js
