+++ src/modules/Auth/components/DecryptPassphrase/DecryptPassphrase.js
@@ -2,16 +2,17 @@
 import React from 'react';
 import { SafeAreaView } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
+import { translate } from 'react-i18next';
+
 import { useTheme } from 'contexts/ThemeContext';
 import HeaderBackButton from 'components/navigation/headerBackButton';
 import { decryptAccount } from 'modules/Auth/utils/decryptAccount';
-import { translate } from 'react-i18next';
 import DropDownHolder from 'utilities/alert';
 import { useAccounts } from 'modules/Accounts/hooks/useAccounts';
-import PasswordForm from '../components/PasswordForm';
-import getStyles from './styles';
+import PasswordForm from '../PasswordForm';
+import getStyles from './DecryptPassphrase.styles';
 
-const DecryptPhrase = ({ account, route, nextStep, t }) => {
+const DecryptPassphrase = ({ account, route, nextStep, t }) => {
   const navigation = useNavigation();
   const { setAccount } = useAccounts();
   const { styles } = useTheme({ styles: getStyles });
@@ -48,4 +49,4 @@ const DecryptPhrase = ({ account, route, nextStep, t }) => {
   );
 };
 
-export default translate()(DecryptPhrase);
+export default translate()(DecryptPassphrase);
