+++ src/modules/Auth/components/PasswordSetupForm/styles.js
