+++ src/modules/Auth/components/PasswordSetupSuccess/styles.js
