+++ src/modules/Auth/components/Register/Register.styles.js
