+++ src/modules/Auth/components/Register/RegisterSuccess/index.js
