+++ src/modules/Auth/components/Register/RegisterSuccess/styles.js
