+++ src/modules/Auth/components/SecretRecoveryPhrase/styles.js
