+++ src/modules/Auth/utils/documentPicker.mock.js
@@ -1,4 +1,4 @@
-import { mockEncryptedAccount } from '../AuthMethod/__fixtures__/mockEncryptedAccount';
+import { mockEncryptedAccount } from '../components/AuthMethod/__fixtures__/mockEncryptedAccount';
 
 export const selectEncryptedFile = async () => {
   return JSON.stringify(mockEncryptedAccount);
