+++ src/modules/Settings/components/BackupPassphrase/BackupPassphrase.js
@@ -4,8 +4,8 @@ import { useRoute } from '@react-navigation/native';
 import { useTheme } from 'contexts/ThemeContext';
 import useScreenshotPrevent from 'hooks/useScreenshotPrevent';
 import Stepper from 'components/shared/Stepper';
-import DecryptPhrase from 'modules/Auth/DecryptPhrase';
-import RegisterConfirm from 'modules/Auth/Register/RegisterConfirm';
+import DecryptPassphrase from 'modules/Auth/components/DecryptPassphrase/DecryptPassphrase';
+import PassphraseQuiz from 'modules/Auth/components/PassphraseQuiz/PassphraseQuiz';
 import { SafeAreaView } from 'react-native-safe-area-context';
 
 import getStyles from './BackupPassphrase.styles';
@@ -16,16 +16,14 @@ export default function BackupPassphrase() {
 
   const { account } = route.params;
 
-  const { styles } = useTheme({
-    styles: getStyles(),
-  });
+  const { styles } = useTheme({ styles: getStyles() });
 
   useScreenshotPrevent();
 
   return (
     <SafeAreaView style={[styles.wrapper, styles.theme.wrapper]}>
       <Stepper currentIndex={0}>
-        <DecryptPhrase
+        <DecryptPassphrase
           account={account}
           route={{
             params: {
@@ -37,7 +35,7 @@ export default function BackupPassphrase() {
 
         <CopyPassphrase />
 
-        <RegisterConfirm customHeader />
+        <PassphraseQuiz customHeader />
       </Stepper>
     </SafeAreaView>
   );
